+++ app/scripts/lizard-nxt-filters.js
@@ -367,3 +367,17 @@ angular.module('lizard-nxt')
 
 });
 
+angular.module('lizard-nxt')
+  .filter('landuseLanduseType', function () {
+  return function (input) {
+    return input.split(' - ')[2];
+  };
+});
+
+angular.module('lizard-nxt')
+  .filter('landuseSource', function () {
+  return function (input) {
+    return input.split(' - ')[1];
+  };
+});
+