+++ app/scripts/directives/search-directive.js
@@ -5,7 +5,8 @@
  * and makes sure the right services are called.
  */
 angular.module('lizard-nxt')
-  .directive('search', ['LocationService', function (LocationService) {
+  .directive('search', ['LocationService', 'ClickFeedbackService',
+  function (LocationService,   ClickFeedbackService) {
 
   var link = function (scope, element, attrs) {
 
@@ -16,7 +17,9 @@ angular.module('lizard-nxt')
      * 13 refers to the RETURN key.
      */
     scope.searchKeyPress = function ($event) {
+
       if ($event.target.id === "searchboxinput") {
+        // Intercept keyPresses *within* searchbox, do xor prevent stuff from happening
         if ($event.which === 13) {
           // User hits [enter] -> do search;
           scope.search();
@@ -35,14 +38,17 @@ angular.module('lizard-nxt')
       if (scope.geoquery && scope.geoquery.length > 1) {
         LocationService.search(scope.geoquery)
           .then(function (response) {
-            scope.cleanInput();
-            scope.box.content.location = {};
-            scope.box.content.location.data = response;
+            scope.geoquery = "";
+            if (response.length !== 0) {
+              scope.box.content.location = {
+                data: response
+              };
+            }
           }
         );
       }
       else {
-        scope.cleanInput();
+        scope.geoquery = "";
       }
     };
 
@@ -57,9 +63,23 @@ angular.module('lizard-nxt')
      * @description resets input field
      * on scope, because also needs to trigger on reset button,
      * not just on succesful search/zoom.
+     *
+     * @description - This does the following:
+     *
+     * (1) - Reset box.type to it's default value, "point";
+     * (2) - Reset the search query to the empty string;
+     * (3) - Reset box.content to an empty object;
+     * (4) - Clear mapState.points arr (used for updating the Url);
+     * (5) - Clear the click feedback.
      */
     scope.cleanInput = function () {
+
+      scope.box.type = "point";
       scope.geoquery = "";
+      scope.box.content = {};
+      scope.mapState.points = [];
+      scope.mapState.here = undefined;
+      ClickFeedbackService.emptyClickLayer(scope.mapState);
     };
 
     /**
@@ -81,8 +101,7 @@ angular.module('lizard-nxt')
       scope.cleanInput();
     };
 
-
-   };
+  };
 
   return {
     link: link,