+++ CHANGES.rst
@@ -28,6 +28,8 @@ Release 0.2b16 (2014-11-26)
 - Timeline does not interfere with initial point/line request with a
   pre-existing layerSlug request
 
+- Timeline shows events on startup.
+
 
 Release 0.2b15 (2014-11-24)
 ---------------------------