+++ CHANGES.rst
@@ -7,6 +7,11 @@ Unreleased ()
 - Escape and x-button in search box reset box, points and remove points from
   url.
 
+- Fix pumpstation and channel new entity types.
+
+- Point clicks now have proper alignment for raster response.
+
+
 Release 0.2b16 (2014-11-26)
 ---------------------------
 
@@ -28,6 +33,8 @@ Release 0.2b16 (2014-11-26)
 
 - Timeline does not interfere with initial point/line request with a
   pre-existing layerSlug request
+
+- Timeline shows events on startup.
 
 
 Release 0.2b15 (2014-11-24)