+++ CHANGES.rst
@@ -4,7 +4,13 @@ Changelog of lizard-nxt client
 Unreleased ()
 -------------
 
--
+- Escape and x-button in search box reset box, points and remove points from
+  url.
+
+- Fix pumpstation and channel new entity types.
+
+- Point clicks now have proper alignment for raster response.
+
 
 Release 0.2b16 (2014-11-26)
 ---------------------------
@@ -27,6 +33,8 @@ Release 0.2b16 (2014-11-26)
 
 - Timeline does not interfere with initial point/line request with a
   pre-existing layerSlug request
+
+- Timeline shows events on startup.
 
 
 Release 0.2b15 (2014-11-24)