+++ CHANGES.rst
@@ -4,8 +4,14 @@ Changelog of lizard-nxt client
 Unreleased ()
 -------------
 
+- IE fix for search bar.
+
 - Escape and x-button in search box reset box, points and remove points from
   url.
+
+
+Release 0.2b17 (2014-11-27)
+---------------------------
 
 - Fix pumpstation and channel new entity types.